refactor: 중복되는 layout 뽑아내기
